test(universal-chart): split tests by feature
Move the catch-all tests into focused feature modules and share their render helpers. Parameterized cases keep the same coverage with less repetition, while a 400-line guard prevents another oversized test module.
